+++ smolyakov.alexander/S1/listIo.hpp
@@ -6,13 +6,13 @@
 
 namespace smolyakov
 {
-  using listOfPairs = smolyakov::List<std::pair<std::string, smolyakov::List<long long>>>;
+  using listOfPairs = smolyakov::List<std::pair<std::string, smolyakov::List<size_t>>>;
   listOfPairs inputList(std::istream& inputStream);
   void outputPairListNames(std::ostream& outputStream, listOfPairs list);
-  void rearrangeValues(listOfPairs& listOfPairs, List<smolyakov::List<long long>>& lists);
-  void outputRearrangedLists(std::ostream& outputStream, List<List<long long>>& lists);
-  void calculateSums(List<List<long long>>& lists, List<long long>& destination);
-  void outputListValues(std::ostream& outputStream, List<long long>& list);
+  void rearrangeValues(listOfPairs& listOfPairs, List<smolyakov::List<size_t>>& lists);
+  void outputRearrangedLists(std::ostream& outputStream, List<List<size_t>>& lists);
+  void calculateSums(List<List<size_t>>& lists, List<size_t>& destination);
+  void outputListValues(std::ostream& outputStream, List<size_t>& list);
 }
 
 #endif
